Refactor: split sentences and spelling quality check features into their own modules. Also constants were extracted and put into their own module.
