+++ nlp_profiler/constants.py
@@ -1,0 +1,9 @@
+DEFAULT_PARALLEL = 'default'
+
+SWIFTER = 'using_swifter'
+
+GRANULAR = 'granular'
+HIGH_LEVEL = 'high_level'
+GRAMMAR_CHECK = 'grammar_check'
+PARALLELISATION_METHOD = 'parallelisation_method'
+NOT_APPLICABLE = "N/A"
