+++ nlp_profiler/core.py
@@ -17,7 +17,6 @@
 ### Jupyter Notebook: https://github.com/neomatrix369/awesome-ai-ml-dl/blob/master/examples/better-nlp/notebooks/jupyter/nlp_profiler.ipynb
 
 import re
-import string
 import tempfile
 from itertools import groupby
 
@@ -28,29 +27,25 @@ import language_tool_python
 import nltk
 import pandas as pd
 from joblib import Parallel, delayed
-# NLP
 from nltk.corpus import stopwords
+# NLP
 from nltk.tokenize import word_tokenize
 # Sentiment Analysis
 from textblob import TextBlob
-from textblob import Word
 from tqdm.auto import tqdm
+import swifter #noqa
 
-DEFAULT_PARALLEL = 'default'
-
-SWIFTER = 'using_swifter'
-
-GRANULAR = 'granular'
-HIGH_LEVEL = 'high_level'
-GRAMMAR_CHECK = 'grammar_check'
-PARALLELISATION_METHOD = 'parallelisation_method'
+from nlp_profiler.constants import \
+    NOT_APPLICABLE, PARALLELISATION_METHOD, DEFAULT_PARALLEL, SWIFTER, \
+    GRANULAR, HIGH_LEVEL, GRAMMAR_CHECK
+from nlp_profiler.sentences import count_sentences
+from nlp_profiler.spelling_quality_check \
+    import spelling_quality_score, spelling_quality, spelling_quality_summarised
 
 nltk.download('stopwords')
 STOP_WORDS = set(stopwords.words('english'))
 
 nltk.download('punkt')
-
-NOT_APPLICABLE = "N/A"
 
 memory = joblib.Memory(tempfile.gettempdir(), compress=9, verbose=0)
 
@@ -312,71 +307,6 @@ def sentiment_subjectivity_score(text: str) -> float:
     return TextBlob(text).sentiment.subjectivity
 
 
-### Spell check
-### See https://en.wikipedia.org/wiki/Words_of_estimative_probability
-### The General Area of Possibility
-spelling_quality_score_to_words_mapping = [
-    ["Very good", 99, 100],  # Very good: Certain: 100%: Give or take 0%
-    ["Quite good", 90, 99],  # Quite Good: Almost Certain: 93%: Give or take 6%
-    ["Good", 87, 90],  # Quite Good: Almost Certain: 93%: Give or take 6%
-    ["Pretty good", 63, 87],  # Pretty: Good: Probable: 75%: Give or take about 12%
-    ["Bad", 40, 63],  # So/so: Chances About Even: 50%: Give or take about 10%
-    ["Pretty bad", 12, 40],  # Pretty bad: Probably Not: 30%: Give or take about 10%
-    ["Quite bad", 2, 12],  # Quite bad: Almost Certainly Not 7%: Give or take about 5%
-    ["Very bad", 0, 2]  # Impossible 0%: Give or take 0%
-]
-
-
-def spelling_quality_summarised(quality: str) -> str:
-    if quality == NOT_APPLICABLE:
-        return NOT_APPLICABLE
-
-    if 'good' in quality.lower():
-        return 'Good'
-
-    return 'Bad'
-
-
-def spelling_quality_score(text: str) -> float:
-    if (not isinstance(text, str)) or (len(text.strip()) == 0):
-        return NOT_APPLICABLE
-
-    tokenized_text = word_tokenize(text.lower())
-
-    misspelt_words_count = 0
-    total_words_checks = 0
-    for each_word in tokenized_text:
-        if each_word not in string.punctuation:
-            spellchecked_word = Word(each_word).spellcheck()
-            _, score = spellchecked_word[0]
-            if score != 1:
-                misspelt_words_count += 1
-            total_words_checks += 1
-    num_of_sentences = count_sentences(text)
-    avg_words_per_sentence = total_words_checks / num_of_sentences
-    result = (avg_words_per_sentence - misspelt_words_count) \
-             / avg_words_per_sentence
-
-    # avg_words_per_sentence = (total_words_checks / num_of_sentences) \
-    #     if num_of_sentences > 0 else 0
-    # result = (avg_words_per_sentence - misspelt_words_count) \
-    #          / avg_words_per_sentence \
-    #     if avg_words_per_sentence > 0 else NOT_APPLICABLE
-    # if result == NOT_APPLICABLE:
-    #     return result
-    return result if result >= 0.0 else 0.0
-
-
-def spelling_quality(score: float) -> str:
-    if score == NOT_APPLICABLE:
-        return NOT_APPLICABLE
-
-    score = float(score) * 100
-    for each_slab in spelling_quality_score_to_words_mapping:  # pragma: no cover
-        if (score >= each_slab[1]) and (score <= each_slab[2]):
-            return each_slab[0]
-
-
 ### Grammar check: this is a very slow process
 ### take a lot of time per text it analysis
 def grammar_check_score(text: str) -> int:
@@ -531,26 +461,6 @@ def count_words(text: str) -> int:
     if not isinstance(text, str):
         return NOT_APPLICABLE
     return len(gather_words(text))
-
-
-### Sentences
-def gather_sentences(text: str) -> list:
-    if not isinstance(text, str):
-        return []
-
-    lines = re.findall(r'([^.]*[^.]*)', text)
-    for index, each in enumerate(lines):
-        if each == '':
-            del lines[index]
-
-    return lines
-
-
-def count_sentences(text: str) -> int:
-    if not isinstance(text, str):
-        return NOT_APPLICABLE
-
-    return len(gather_sentences(text))
 
 
 ### Number of spaces
