+++ tests/granular/test_alphanumeric.py
@@ -1,7 +1,9 @@
-from nlp_profiler.core \
-    import NOT_APPLICABLE, gather_alpha_numeric, count_alpha_numeric  # noqa
 import numpy as np
 import pytest
+
+from nlp_profiler.constants import NOT_APPLICABLE
+from nlp_profiler.core \
+    import gather_alpha_numeric, count_alpha_numeric  # noqa
 
 text_with_alphanumeric_chars = '2833047 people live in this area'
 
@@ -23,6 +25,7 @@ def test_given_invalid_text_when_parsed_then_return_empty_list(
     # then
     assert expected_result == actual_result, \
         f"Expected: {expected_result}, Actual: {actual_result}"
+
 
 text_to_return_count_mapping = [
     (np.nan, NOT_APPLICABLE),
