+++ tests/granular/test_chars_and_spaces.py
@@ -1,5 +1,6 @@
+from nlp_profiler.constants import NOT_APPLICABLE
 from nlp_profiler.core \
-    import NOT_APPLICABLE, count_chars, count_spaces, count_characters_excluding_spaces  # noqa
+    import count_chars, count_spaces, count_characters_excluding_spaces  # noqa
 import numpy as np
 import pytest
