+++ tests/granular/test_non_alphanumeric.py
@@ -1,7 +1,9 @@
-from nlp_profiler.core \
-    import NOT_APPLICABLE, gather_non_alpha_numeric, count_non_alpha_numeric  # noqa
 import numpy as np
 import pytest
+
+from nlp_profiler.constants import NOT_APPLICABLE
+from nlp_profiler.core \
+    import gather_non_alpha_numeric, count_non_alpha_numeric  # noqa
 
 text_with_a_number = '2833047 people live in this area'
 text_with_emojis = "I love ⚽ very much 😁"
