+++ tests/granular/test_sentences.py
@@ -1,7 +1,8 @@
-from nlp_profiler.core \
-    import NOT_APPLICABLE, gather_sentences, count_sentences  # noqa
 import numpy as np
 import pytest
+
+from nlp_profiler.constants import NOT_APPLICABLE
+from nlp_profiler.sentences import gather_sentences, count_sentences  # noqa
 
 text_with_emojis = "I love ⚽ very much 😁"
 text_with_a_number = '2833047 people live in this area'
