+++ tests/high_level/test_spelling_check.py
@@ -3,7 +3,8 @@ import math
 import numpy as np
 import pytest
 
-from nlp_profiler.core import NOT_APPLICABLE, spelling_quality_score, \
+from nlp_profiler.constants import NOT_APPLICABLE
+from nlp_profiler.spelling_quality_check import spelling_quality_score, \
     spelling_quality, spelling_quality_summarised  # noqa
 
 good_spelling_text = 'People live in this area. It is not a good area. People live in this area. It is not a good area. 2833047 people live in this area. It is not a good area.'
